[CMake] fix component requirement
In #143, the QtAll component of PythonQt is required, meaning that if we
find PythonQt but not its QtAll component, we should consider that
PythonQt was not found.

I thought that `FIND_PACKAGE(PythonQt COMPONENTS QtAll REQUIRED)` as
opposed to `FIND_PACKAGE(PythonQt REQUIRED COMPONENTS QtAll REQUIRED)`
would mean that, but actually, this second `REQUIRED` implies a failure
of the configure step if PythonQt is not found.